requirements.txt
streamlit
PyPDF2
python-dotenv
openai
langchain
langchain-community
faiss-cpu
tiktoken